Move IDocumentSnapshot, IProjectSnapshot, DocumentContext and ISolutionQueryOperations, up to Remote.Razor
