+++ src/Razor/src/Razor/src/Microsoft.CodeAnalysis.Remote.Razor/Tooltip/IComponentAvailabilityService.cs
@@ -4,7 +4,7 @@
 using System.Collections.Immutable;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.CodeAnalysis.Razor.ProjectSystem;
+using Microsoft.CodeAnalysis.Remote.Razor.ProjectSystem;
 
 namespace Microsoft.CodeAnalysis.Remote.Razor.Tooltip;
