+++ src/Razor/src/Razor/test/Microsoft.CodeAnalysis.Remote.Razor.UnitTests/Tooltip/TestComponentAvailabilityService.cs
@@ -5,7 +5,7 @@ using System.Collections.Immutable;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Razor.Threading;
-using Microsoft.CodeAnalysis.Razor.ProjectSystem;
+using Microsoft.CodeAnalysis.Remote.Razor.ProjectSystem;
 
 namespace Microsoft.CodeAnalysis.Remote.Razor.Tooltip;
